backend/requirements.txt
fastapi
uvicorn
neo4j
python-dotenv
pydantic

# AI & Ingestion
llama-parse
llama-index-core
llama-index-llms-groq
llama-index-llms-openai